Prevent losing images on products when archiving the original
product_images table had a foreign key reference to blobs. Blobs had
a variant column, included in the primary key. blobs was split in two:
one for the reference, one for the variants. The variants table stores
the actual variant & data. The metadata is stored in blobs. This
allows a clean split and allows product_images to keep a strong foreign
key to blobs, but this time we won't lose the image when we rescale and
archive the original.